Add GCC argument checking for printf-format strings, fix warnings: ../src/btree/bt_discard.c: In function ‘__wt_page_free’: ../src/btree/bt_discard.c:30: warning: implicit conversion shortens 64-bit value into a 32-bit value ../src/btree/bt_discard.c:30: warning: implicit conversion shortens 64-bit value into a 32-bit value

+++ src/api/api.c
@@ -125,7 +125,7 @@ __session_create(WT_SESSION *wt_session,
 		key_format = "u";
 	else {
 		__wt_errx(session, "Unknown key_format '%.*s'",
-		    cval.len, cval.str);
+		    (int)cval.len, cval.str);
 		return (EINVAL);
 	}
 
@@ -139,7 +139,7 @@ __session_create(WT_SESSION *wt_session,
 		value_format = "u";
 	} else {
 		__wt_errx(session, "Unknown value_format '%.*s'",
-		    cval.len, cval.str);
+		    (int)cval.len, cval.str);
 		return (EINVAL);
 	}
